Make pry dumber
We are failing to explain the difference between a collection and a record properly, leading to millions of questions [like this](https://piazza.com/class/jxc5bt9wqba28c?cid=36); even from students with software development experience.

This patch adds a `.pryrc` configuration file that overrides the default representation of `ActiveRecord::Relation` to make it dumber/less informative/hopefully clearer to students. It will only display `Array of 323 Photo`, and force them to use `at` to go further.